edge_model_serving/onnx_serving/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
numpy>=1.25.0
Pillow>=10.0.0
onnxruntime>=1.15.0
pytest>=7.0.0
requests>=2.31.0
